requirements.txt
aiohttp==3.8.1
aiohttp-apispec==2.2.3
pydantic==1.10.2
aiohttp-pydantic==1.12.2
aiohttp-swagger==1.0.16
PyYAML~=6.0
multidict~=6.0.3
sqlalchemy==1.4.40
alembic==1.8.0
psycopg2-binary
asyncpg==0.26.0
aiohttp_middlewares==2.1.0
aiohttp-session==2.12.0
cryptography==38.0.4
pyjwt==2.6.0
pytz==2022.6
pytest==7.2.1
pytest-aiohttp==1.0.4
pytest-asyncio==0.20.3
frozenlist==1.3.3
pre-commit==3.0.0
isort==5.11.4
coverage==7.1.0

